% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.R
\name{bsp}
\alias{bsp}
\title{Define a Beta-Stacy Process}
\usage{
bsp(support, centeringMeasure, precision, calculateMoments = FALSE)
}
\arguments{
\item{support}{A numeric vector indicating the points at which the support for the BSP is defined}

\item{centeringMeasure}{The mean or centering measure of the BSP at each point on the support
or a function that can be evaluated at each point}

\item{precision}{A constant indicating the precision of the BSP (Currently, the precision
cannot vary over the support).}

\item{calculateMoments}{This is used to calculate the second moment while
creating the bsp. The moment is used when merging components with other
components. If this function is only being used to create a prior, it can
normally be set to false.}
}
\value{
An object representing a Beta-Stacy Process.
}
\description{
Define a Beta-Stacy Process
}
\details{
The most common use case for this function will be defining BSPs to use as priors.
The precision reflects the uncertainty in the prior. In general, a one unit increase
in the precision is equivalent to one additional observation.
}
\examples{
bsp(support=c(1,5), centeringMeasure=c(0.2,0.4), precision=2)



}
