% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspPosterior.R
\name{bsp_sampling}
\alias{bsp_sampling}
\title{Creates a matrix where each column represents a sampled CDF from the provided BSP}
\usage{
bsp_sampling(BSP, reps = 10000)
}
\arguments{
\item{BSP}{Any BSP object}

\item{reps}{Number of samples to draw, default 10000}
}
\value{
A matrix. Each column can be plotted against the BSP's support to view a sampled CDF from the BSP
}
\description{
Creates a matrix where each column represents a sampled CDF from the provided BSP
}
