% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{bspConfint}
\alias{bspConfint}
\title{Finds approximate (simulated) confidence intervals for the percent failed
at a given time}
\usage{
bspConfint(bsp, times, alpha = 0.05)
}
\arguments{
\item{bsp}{The bsp object}

\item{times}{Place where you would like confidence intervals}
}
\value{
A 2xlength(times) matrix with upper and lower bounds of the confidence interval
}
\description{
Finds approximate (simulated) confidence intervals for the percent failed
at a given time
}
\note{
If you intend to call this function multiple times on the same object,
calling bsp$Samples<-bspSampling(bsp) before bspConfint
will increase performance.
}
\examples{
bsp=bsp(c(1:3), centeringMeasure = c(.1,.9, .98), precision = 2)
bspConfint(bsp, 1:3)

}
