% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{bspConfint2}
\alias{bspConfint2}
\title{Finds approximate confidence interval by fitting a beta distribution to the BSP moemnts}
\usage{
bspConfint2(bsp, times, conf.level = 0.05)
}
\arguments{
\item{bsp}{The bsp object}

\item{times}{Place where you would like confidence intervals}

\item{conf.level}{Confidence level desired (default .05)}
}
\value{
A 2xlength(times) matrix with upper and lower bounds of the confidence interval
}
\description{
Finds approximate confidence interval by fitting a beta distribution to the BSP moemnts
}
\examples{
bsp=bsp(c(1:3), centeringMeasure = c(.1,.9, .98), precision = 2)
bspConfint2(bsp, 1:3)

}
