% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspMoments.R
\name{bspFromMoments}
\alias{bspFromMoments}
\title{Creates a bsp with given moments at each spot on the support}
\usage{
bspFromMoments(E1, E2, support)
}
\arguments{
\item{E1}{The first moment}

\item{E2}{The second moment}

\item{support}{The times for the first and second moments (all vectors equal length)}
}
\value{
A bsp object with those moments
}
\description{
Creates a bsp with given moments at each spot on the support
}
\examples{
bsp=bspFromMoments(c(.2, .4,.8), c(.5, .5, .5), 1:3)
bsp


}
