% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{bspSamples}
\alias{bspSamples}
\title{Obtains realizations of a bsp}
\usage{
bspSamples(bsp, reps = 10000)
}
\arguments{
\item{bsp}{The bsp object}

\item{reps}{Number of realizations to return}
}
\value{
Returns an length(support) x reps matrix. Each column corresponds to a valid discrete CDF drawn from the bsp.
}
\description{
Obtains realizations of a bsp
}
\examples{
bsp=bsp(c(1:3), centeringMeasure = c(.1,.9, .98), precision = 2)
samples=bspSamples(bsp)
#To visualize
matplot(samples, bsp$support, type='l')


}
