% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{bspSampling}
\alias{bspSampling}
\title{Creates a matrix where each column represents a sampled CDF from the provided BSP}
\usage{
bspSampling(bsp, reps = 10000)
}
\arguments{
\item{bsp}{Any bsp object}

\item{reps}{Number of samples to draw, default 10000}
}
\value{
A matrix. Each column can be plotted against
the BSP's support to view a sampled CDF from the BSP.
Rownames are the support point for that row. Use samplesAt to
get extract sampled distribution at single point
}
\description{
Creates a matrix where each column represents a sampled CDF from the provided BSP
}
\examples{
bsp=bsp(c(1:3), centeringMeasure = c(.1,.9, .98), precision = 2)
samples=bspSamples(bsp)
#To visualize
matplot(samples, bsp$support, type='l')

}
