% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityFileParsing.R
\name{createUninformativePrior}
\alias{createUninformativePrior}
\title{Creates a non-informative prior}
\usage{
createUninformativePrior(data)
}
\arguments{
\item{data}{The data matrix, which is used to ensure only the points necessary appear on the support}
}
\value{
A bsp object with 0 precision.
}
\description{
Creates a non-informative prior
}
