% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspMoments.R
\name{E1E2}
\alias{E1E2}
\title{Calculates the first and second moments of a bsp object at each spot on the support}
\usage{
E1E2(bsp)
}
\arguments{
\item{bsp}{The bsp object}
}
\value{
Vector of second moments that aligns with bsp$support
}
\description{
Calculates the first and second moments of a bsp object at each spot on the support
}
\examples{
bsp=bsp(c(1:3), centeringMeasure = c(.1,.9, .98), precision = 2)
bsp$E2=E1E2(bsp)
bsp$E2

}
