% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspMoments.R
\name{E1E2_parallel}
\alias{E1E2_parallel}
\title{Calculates the first and second moments of the subsystem if bsp1 and bsp2 are in parallel}
\usage{
E1E2_parallel(bsp1, bsp2)
}
\arguments{
\item{bsp1}{The first bsp object}

\item{bsp2}{The second bsp object}
}
\value{
A list with three elements, all equal length, containing the
support, E1, and E2 for the new bsp
}
\description{
Calculates the first and second moments of the subsystem if bsp1 and bsp2 are in parallel
}
