% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityFileParsing.R
\name{estimateSystemReliability}
\alias{estimateSystemReliability}
\title{Estimates system reliability for a specified system.}
\usage{
estimateSystemReliability(file, priorList, dataList)
}
\arguments{
\item{file}{A string with the file path containing the reliability diagram (see details)}

\item{priorList}{A list with named elements containing bsp object representing the prior for each component in the reliability diagram}

\item{dataList}{A list with named elements for each component or subsystem for which reliability was measured}
}
\value{
A named list of posterior bsp objects, one for each named component in the system
(if no data was provided the posterior will be equivalent to the provided prior)
}
\description{
Estimates system reliability for a specified system.
}
\section{file Details}{

The file is a text file specifying your reliability block diagram. Parallel and series subsystems can be specified as in the following example:

S(Engine, GasDelivery):GasPropulsion

S(Motor, Batteries, Controller, Belt):ElectricPropulsion

P(GasPropulsion, ElectricPropulsion):Propulsion

S and P represent series and parallel relationships respectively.
The first line of this diagram can be read as
"Engine and GasDelivery are components in series
of a subsystem named GasPropulsion." For each component on the right hand
side of an equation, the prior of that component will be determined
by the posterior of its subcomponents. Data, if available, can be used to update its
posterior by adding the data to dataList (e.g. assigning a matrix to dataList$GasPropulsion).
For each component on the left hand side of the colon, priors can be defined
by adding a bsp object to priorList. If no prior is provided, a non-informative
prior will be used, with a warning.

The following syntax conventions for the file will be enforced:
\itemize{
\item Each line begins with S( or P( indicating series or parallel, one relationship per line

\item Components can be uppercase, lowercase, and include numbers, but no special characters.

\item All relationships must be named by adding a colon and a label after the end of the relationship

\item Two or more elements can be specified in a relationship

\item A check will be done to ensure there are no circular dependencies among the named components.
For example, Propulsion cannot be used as an input to GasPropulsion because GasPropulsion
is an input to Propulsion
}
}

\section{priorList Details}{

The priorList contains the bsp prior objects used as objects for the components.
For example, if "valve" is a component appearing only on the left hand side of a relationship
in the reliability diagram, priorList$valve must contain a betaStaceyProcess object
representing the prior. These objects can be specified with the bsp() function.
Objects appearing on the right hand side of a relationship do not need a prior, as
their prior will be computed from the posteriors of the component pieces.
If a prior is not provided for a LHS component, a non-informative prior will be
created and used, with a warning.
}

\section{dataList Details}{

A dataList element should be provided for any component on which data has been gathered.
Each dataList element (eg dataList$valve) should be a nx2 matrix with the first column
being the observed failurer time (or end or test) and the second column contains the censoring variable
0 - if right censored 1 - if fully observed.

Any object for which data is provided will have a new bsp object computed as the
posterior of the prior and the provided data.
}

\examples{
file="ReliabilityDiagram.txt"
write.table("S(Engine, GasDelivery):GasPropulsion", file=file, quote=F, row.names=F, col.names=F)
priorList<-list(Engine = bsp(1, .5, 1), GasDelivery=bsp(3,.6, 1))
dataList<-list(Engine=matrix(c(2, 1, 3, 0), byrow=T, nrow=2),
GasDelivery=matrix(c(5, 1, 6, 1), byrow=T, nrow=2))
estimateSystemReliability(file, priorList, dataList)
}
