% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.R
\name{evaluate_centering_measures}
\alias{evaluate_centering_measures}
\title{Evaluate centering measure at specific times}
\usage{
evaluate_centering_measures(bsp, times)
}
\arguments{
\item{bsp}{A BSP object}

\item{times}{A list of times that the centering measure should be determined for}
}
\value{
A vector of same length as times with the centering measure for each time
}
\description{
Evaluate centering measure at specific times
}
\note{
For times in between jumps on the support, the centering measure is
considered equal to its value after the last jump
}
\examples{
evaluate_centering_measure(bsp(c(1,2), c(.2,.6), 1), c(.5,1.5,2.5))
}
