% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.R
\name{evaluate_precision}
\alias{evaluate_precision}
\title{Evaluate precision at specific times}
\usage{
evaluate_precision(bsp, times)
}
\arguments{
\item{bsp}{A BSP object}

\item{times}{A list of times that the centering measure should be determined for}
}
\value{
A vector of same length as times with the centering measure for each time
}
\description{
Evaluate precision at specific times
}
\examples{
evaluate_precision(bsp(c(1,2), c(.2,.6), 1), c(.5,1.5,2.5))
}
