% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{print.betaStacyProcess}
\alias{print.betaStacyProcess}
\title{Prints a Beta-Stacy process object as a data frame}
\usage{
\method{print}{betaStacyProcess}(x)
}
\arguments{
\item{x}{The bsp object to be printed}
}
\value{
none
}
\description{
Prints a Beta-Stacy process object as a data frame
}
\examples{
print(bsp(c(1,5),c(0.2,0.4),c(2,1)))


}
