% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{quantile.betaStacyProcess}
\alias{quantile.betaStacyProcess}
\title{Calculates the expected time to failure for various quantiles}
\usage{
\method{quantile}{betaStacyProcess}(bsp, probs)
}
\arguments{
\item{bsp}{The bsp object to calculate from}

\item{probs}{numeric vector of probabilities}
}
\value{
a named numeric vector of times associated with each quantile
}
\description{
Calculates the expected time to failure for various quantiles
}
\details{
This function calculates the expected failure time for a given quantile from
the centering measure of the BSP. Linear interpolation is to choose time between two discrete points
}
\examples{
bsp=bsp(1:3, c(.25,.5,.75), 1)
quantile(bsp, c(.333, .666))
}
