% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{samplesAt}
\alias{samplesAt}
\title{Extract distribution at select point}
\usage{
samplesAt(samples, time)
}
\arguments{
\item{samples}{The samples produced by bspSampling()}

\item{time}{The spot on the support at which you would like to examine the distribution}
}
\value{
A vector of samples drawn from that point on the support
}
\description{
Extract distribution at select point
}
\examples{
bsp=bsp(1:3, c(.2,.5,.8), 3)
samples=bspSampling(bsp)
marginal = sampleAt(samples, 2.5)
}
