% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspMoments.R
\name{bspFromMoments}
\alias{bspFromMoments}
\title{Creates a bsp with given moments at each spot on the support}
\usage{
bspFromMoments(mList)
}
\arguments{
\item{mList}{A list with three elements, all equal length, containing the
support, E1, and E2 for the new bsp}
}
\value{
A bsp object with those moments
}
\description{
Creates a bsp with given moments at each spot on the support
}
\examples{
bsp=bspFromMoments(list(E1=c(.2, .4,.8), E2=c(.5, .5, .5), support=1:3))
bsp


}
