% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspPosterior.R
\name{bspPosterior}
\alias{bspPosterior}
\title{Update a BSP prior with data and return a BSP posterior}
\usage{
bspPosterior(bspPriorObject, data)
}
\arguments{
\item{data}{An nx2 matrix of data the first column are the actual data
the second column contains the censoring variable
0 - if right censored 1 - if fully observed}

\item{bspPrior}{The BSP object that defines the BSP prior}
}
\value{
An object representing a Beta-Stacy Process.
}
\description{
Update a BSP prior with data and return a BSP posterior
}
\examples{
bsp(c(1,5),c(0.2,0.4),c(2,1))


}
